% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSPS.R
\name{getSPS}
\alias{getSPS}
\title{Generate set of stable phosphoporylated sites}
\usage{
getSPS(phosData, conds)
}
\arguments{
\item{phosData}{a list of users' PhosphoExperiment objects from which generate SPSs}

\item{conds}{a list of vector contains the conditions labels for each sample in the phosphoExperiment objects}

\item{num}{the number of identified SPSs, by default is 100}

\item{residueInfo}{whether the phosphosite contains amino acid information or not, by default is FALSE}
}
\value{
A vectors of stably phosphorylated sites
}
\description{
Generate set of stable phosphoporylated sites
}
\examples{

library(stringr)

data("phospho_L6_ratio_pe")
data("phospho_liverInsTC_RUV_pe")

ppe1 <- phospho.L6.ratio.pe
ppe2 <- phospho.liver.Ins.TC.ratio.RUV.pe

ppe.list <- list(ppe1, ppe2)
cond.list <- list(grp1 = gsub("_.+", "", colnames(ppe1)),
                  grp2 = stringr::str_sub(colnames(ppe2), end=-5))

inhouse_SPSs <- getSPS(ppe.list, conds = cond.list, residueInfo = FALSE)

}
