% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hSPSs}
\alias{hSPSs}
\title{A list of human Stably Phosphorylated Sites (hSPSs)}
\format{
An object of class \code{character} of length 326.
}
\usage{
data(hSPSs)
}
\description{
A list of stably phosphoryalted sites defined from a panel of
human phosphoproteomics datasets.
}
\keyword{datasets}
