% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KinaseFamily}
\alias{KinaseFamily}
\title{KinaseFamily}
\format{
An object of class \code{matrix} (inherits from \code{array})
with 425 rows and 6 columns.
}
\usage{
data(KinaseFamily)
}
\description{
A summary table of kinase family
}
\keyword{datasets}
