% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif.human.list}
\alias{motif.human.list}
\title{List of human kinase motifs}
\format{
An object of class \code{list} of length 380.
}
\usage{
data(KinaseMotifs)
}
\description{
A list of human kinase motifs and their sequence probability
matrix.
}
\keyword{datasets}
