% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif.mouse.list}
\alias{motif.mouse.list}
\title{List of mouse kinase motifs}
\format{
An object of class \code{list} of length 250.
}
\usage{
data(KinaseMotifs)
}
\description{
A list of mouse kinase motifs and their sequence probability
matrix.
}
\keyword{datasets}
