% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif.rat.list}
\alias{motif.rat.list}
\title{List of rat kinase motifs}
\format{
An object of class \code{list} of length 159.
}
\usage{
data(KinaseMotifs)
}
\description{
A list of rat kinase motifs and their sequence probability
matrix.
}
\keyword{datasets}
