% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mSEGs}
\alias{mSEGs}
\title{A list of Stably Expressed Genes (SEGs)}
\format{
An object of class \code{character} of length 916.
}
\usage{
data(hSEGs)
}
\description{
A list of stably expressed genes (SEGs) in mouse and human 
identified from a collection of single-cell RNA-sequencing data. See Lin et al., 
Evaluating stably expressed genes in single cells, GigaScience, 8(9):giz106, 
https://doi.org/10.1093/gigascience/giz106 for more details
}
\keyword{datasets}
