% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pathways.KEGG}
\alias{Pathways.KEGG}
\title{KEGG pathway annotations}
\format{
An object of class \code{list} of length 186.
}
\usage{
data(Pathways)
}
\description{
The data object contains the annotations of KEGG pathways.
}
\keyword{datasets}
