% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pathways.reactome}
\alias{Pathways.reactome}
\title{Reactome pathway annotations}
\format{
An object of class \code{list} of length 674.
}
\usage{
data(Pathways)
}
\description{
The data object contains the annotations of Reactome pathways.
}
\keyword{datasets}
