% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phospho.cells.Ins}
\alias{phospho.cells.Ins}
\alias{phospho.cells.Ins.pe}
\title{phospho.cells.Ins}
\format{
An object of class \code{matrix} (inherits from \code{array})
with 49617 rows and 24 columns.

An object of class \code{matrix} (inherits from \code{array})
with 49617 rows and 24 columns.
}
\source{
doi: 10.1038/nbt.3327 (PXD001792)

doi: 10.1038/nbt.3327 (PXD001792)
}
\usage{
data(phospho.cells.Ins.sample)

data(phospho.cells.Ins.pe)
}
\description{
A subset of phosphoproteomics dataset generated by 
Humphrey et al., [doi:10.1038/nbt.3327] from two mouse liver cell lines 
(Hepa1.6 and FL38B) that were treated with either PBS (mock) or insulin.

A phosphoproteome Object containing a subset of phosphoproteomics 
dataset generated by Humphrey et al., [doi:10.1038/nbt.3327] from two mouse 
liver cell lines (Hepa1.6 and FL38B) that were treated with either PBS (mock) 
or insulin.
}
\references{
Humphrey et al., 2015, doi: 10.1038/nbt.3327

Humphrey et al., 2015, doi: 10.1038/nbt.3327
}
\keyword{datasets}
