% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phospho.L6.ratio.pe}
\alias{phospho.L6.ratio.pe}
\title{phospho_L6_ratio_pe}
\format{
An PhosphoExperiment object
}
\source{
PRIDE accesion number: PXD001792
}
\usage{
data(phospho_L6_ratio_pe)
}
\description{
L6 myotube phosphoproteome dataset
(accession number: PXD019127).
}
\keyword{datasets}
