% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phospho.L6.ratio}
\alias{phospho.L6.ratio}
\title{phospho.L6.ratio}
\format{
An object of class \code{matrix} (inherits from \code{array})
with 6660 rows and 12 columns.
}
\source{
PRIDE accesion number: PXD001792
}
\usage{
data(phospho_L6_ratio)
}
\description{
An L6 myotube phosphoproteome dataset
(accession number: PXD019127).
}
\keyword{datasets}
