% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phospho.liver.Ins.TC.ratio.RUV.pe}
\alias{phospho.liver.Ins.TC.ratio.RUV.pe}
\title{phospho.liver.Ins.TC.ratio.RUV.pe}
\format{
A Phosphoproteome Object
}
\source{
PRIDE accesion number: PXD001792
}
\usage{
data(phospho.liver.Ins.TC.ratio.RUV.pe)
}
\description{
A subset of phosphoproteomics dataset integrated from two
time-course datasets of early and intermediate insulin signalling in mouse
liver upon insulin stimulation.
}
\references{
Humphrey et al., 2015
}
\keyword{datasets}
