% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phospho.liver.Ins.TC.ratio.RUV}
\alias{phospho.liver.Ins.TC.ratio.RUV}
\title{phospho_liverInsTC_RUV_sample}
\format{
An object of class \code{matrix} (inherits from \code{array})
with 5000 rows and 90 columns.
}
\source{
PRIDE accesion number: PXD001792
}
\usage{
data(phospho_liverInsTC_RUV_sample)
}
\description{
A subset of phosphoproteomics dataset integrated from two
time-course datasets of early and intermediate insulin signalling in mouse
liver upon insulin stimulation.
}
\references{
Humphrey et al., 2015
}
\keyword{datasets}
