% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PhosphoSite.rat}
\alias{PhosphoSite.rat}
\title{PhosphoSitePlus annotations for rat}
\format{
An object of class \code{list} of length 158.
}
\source{
\url{https://www.phosphosite.org}
}
\usage{
data(PhosphoSitePlus)
}
\description{
The data object contains the annotations of kinases and their
conrresponding substrates as phosphorylation sites in rat.
It is extracted from the PhosphoSitePlus database.
For details of PhosphoSitePlus, please refer to the article:
Hornbeck et al. Nucleic Acids Res. 40:D261-70, 2012
}
\keyword{datasets}
