% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinaseSubstratePrediction.R
\name{siteAnnotate}
\alias{siteAnnotate}
\title{Phosphosite annotation}
\usage{
siteAnnotate(site, phosScoringMatrices, predMatrix)
}
\arguments{
\item{site}{site the ID of a phosphosite}

\item{phosScoringMatrices}{output from function kinaseSubstrateScore()}

\item{predMatrix}{a prediction matrix from kinaseSubstratePred()}
}
\value{
A graphical plot
}
\description{
This function plots the combined scores of each of all kinases
for a given phosphosites
}
\examples{

data('phospho.L6.ratio.pe')
data('SPSs')

data('phospho.L6.ratio.pe')
data('SPSs')
data('PhosphoSitePlus')

grps = gsub('_.+', '', colnames(phospho.L6.ratio.pe))

# Construct a design matrix by condition
design = model.matrix(~ grps - 1)

# phosphoproteomics data normalisation using RUV
L6.sites = paste(sapply(phospho.L6.ratio.pe@GeneSymbol, function(x)paste(x)),
                 ";",
                 sapply(phospho.L6.ratio.pe@Residue, function(x)paste(x)),
                 sapply(phospho.L6.ratio.pe@Site, function(x)paste(x)),
                 ";", sep = "")
ctl = which(L6.sites \%in\% SPSs)
phospho.L6.ratio.RUV = RUVphospho(phospho.L6.ratio.pe@assays@data$Quantification, 
                                  M = design, k = 3,ctl = ctl)

phosphoL6 = phospho.L6.ratio.RUV
rownames(phosphoL6) <- rownames(phospho.L6.ratio.pe)

# filter for up-regulated phosphosites
phosphoL6.mean <- meanAbundance(phosphoL6, grps = grps)
aov <- matANOVA(mat=phosphoL6, grps = grps)
phosphoL6.reg <- phosphoL6[(aov < 0.05) &
                        (rowSums(phosphoL6.mean > 0.5) > 0), drop = FALSE]
L6.phos.std <- standardise(phosphoL6.reg)

idx <- match(rownames(L6.phos.std), L6.sites)
L6.phos.seq <- phospho.L6.ratio.pe@Sequence[idx]

L6.matrices <- kinaseSubstrateScore(PhosphoSite.mouse, L6.phos.std,
    L6.phos.seq, numMotif = 5, numSub = 1)
set.seed(1)
L6.predMat <- kinaseSubstratePred(L6.matrices, top=30)
dev.off()

# We will look at the phosphosite AAK1;S677 for demonstration purpose.
site = "AAK1;S677;"
siteAnnotate(site, L6.matrices, L6.predMat)

}
