% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPSs}
\alias{SPSs}
\title{A list of Stably Phosphorylated Sites (SPSs)}
\format{
An object of class \code{character} of length 100.
}
\usage{
data(SPSs)
}
\description{
A list of stably phosphoryalted sites defined from a panel of
phosphoproteomics datasets. For full list of the datasets used,
please refer to our preprint for the full list.
}
\keyword{datasets}
