% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinaseSubstratePrediction.R
\name{kinaseSubstrateHeatmap}
\alias{kinaseSubstrateHeatmap}
\title{Kinase-substrate annotation prioritisation heatmap}
\usage{
kinaseSubstrateHeatmap(phosScoringMatrices, top = 3)
}
\arguments{
\item{phosScoringMatrices}{a matrix returned from kinaseSubstrateScore.}

\item{top}{the number of top ranked phosphosites for each kinase to be
included in the heatmap. Default is 1.}
}
\value{
a pheatmap object.
}
\description{
Kinase-substrate annotation prioritisation heatmap
}
\examples{
data('phospho_L6_ratio_pe')
data('SPSs')
data('PhosphoSitePlus')

ppe <- phospho.L6.ratio.pe
sites = paste(sapply(ppe@GeneSymbol, function(x)x),";",
    sapply(ppe@Residue, function(x)x),
    sapply(ppe@Site, function(x)x),
    ";", sep = "")
grps = gsub("_.+", "", colnames(ppe))
design = model.matrix(~ grps - 1)
ctl = which(sites \%in\% SPSs)
ppe = RUVphospho(ppe, M = design, k = 3, ctl = ctl)

phosphoL6 = ppe@assays@data$normalised

# filter for up-regulated phosphosites
phosphoL6.mean <- meanAbundance(phosphoL6, grps = grps)
aov <- matANOVA(mat=phosphoL6, grps = grps)
idx <- (aov < 0.05) & (rowSums(phosphoL6.mean > 0.5) > 0)
phosphoL6.reg <- phosphoL6[idx, ,drop = FALSE]

L6.phos.std <- standardise(phosphoL6.reg)

rownames(L6.phos.std) <- paste0(ppe@GeneSymbol, ";", ppe@Residue, ppe@Site, ";")[idx]

L6.phos.seq <- ppe@Sequence[idx]

L6.matrices <- kinaseSubstrateScore(PhosphoSite.mouse, L6.phos.std,
    L6.phos.seq, numMotif = 5, numSub = 1)
    
kinaseSubstrateHeatmap(L6.matrices)

}
