% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinaseSubstratePrediction.R
\name{kinaseSubstrateScore}
\alias{kinaseSubstrateScore}
\title{Kinase substrate scoring}
\usage{
kinaseSubstrateScore(
    substrate.list,
    mat,
    seqs,
    numMotif = 5, 
    numSub = 1, 
    verbose = TRUE
)
}
\arguments{
\item{substrate.list}{a list of kinases with each element containing an array
of substrates.}

\item{mat}{a matrix with rows correspond to phosphosites and columns
correspond to samples.}

\item{seqs}{an array containing aa sequences surrounding each of all
phosphosites.
Each sequence has length of 15 (-7, p, +7).}

\item{numMotif}{minimum number of sequences used for compiling motif for
each kinase.
Default is 5.}

\item{numSub}{minimum number of phosphosites used for compiling
phosphorylation
profile for each kinase. Default is 1.}

\item{verbose}{Default to \code{TRUE} to show messages during the progress.
All messages will be suppressed if set to \code{FALSE}}
}
\value{
A list of 4 elements.
\code{motifScoreMatrix}, \code{profileScoreMatrix},
\code{combinedScoreMatrix}, \code{ksActivityMatrix} (kinase activity matrix)
and their \code{weights}.
}
\description{
This function generates substrate scores for kinases that pass
filtering based on both motifs and dynamic profiles
}
\examples{

data('phospho_L6_ratio_pe')
data('SPSs')
data('PhosphoSitePlus')

ppe <- phospho.L6.ratio.pe
sites = paste(sapply(ppe@GeneSymbol, function(x)x),";",
    sapply(ppe@Residue, function(x)x),
    sapply(ppe@Site, function(x)x),
    ";", sep = "")
grps = gsub("_.+", "", colnames(ppe))
design = model.matrix(~ grps - 1)
ctl = which(sites \%in\% SPSs)
ppe = RUVphospho(ppe, M = design, k = 3, ctl = ctl)

phosphoL6 = ppe@assays@data$normalised

# filter for up-regulated phosphosites
phosphoL6.mean <- meanAbundance(phosphoL6, grps = grps)
aov <- matANOVA(mat=phosphoL6, grps = grps)
idx <- (aov < 0.05) & (rowSums(phosphoL6.mean > 0.5) > 0)
phosphoL6.reg <- phosphoL6[idx, ,drop = FALSE]

L6.phos.std <- standardise(phosphoL6.reg)

rownames(L6.phos.std) <- paste0(ppe@GeneSymbol, ";", ppe@Residue, ppe@Site, ";")[idx]

L6.phos.seq <- ppe@Sequence[idx]

L6.matrices <- kinaseSubstrateScore(PhosphoSite.mouse, L6.phos.std,
    L6.phos.seq, numMotif = 5, numSub = 1)

}
