% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhosphoExperiment.R
\docType{class}
\name{PhosphoExperiment-class}
\alias{PhosphoExperiment-class}
\alias{PhosphoExperiment}
\alias{coerce,}
\alias{SummarizedExperiment,}
\alias{PhosphoExperiment-method}
\title{The PhosphoExperiment class}
\usage{
PhosphoExperiment(
  ...,
  UniprotID = c(),
  GeneSymbol = c(),
  Site = c(),
  Residue = c(),
  Sequence = c()
)
}
\description{
The PhosphoExperiment class
}
