% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhosphoExperiment.R
\docType{class}
\name{PhosphoExperiment-class}
\alias{PhosphoExperiment-class}
\alias{PhosphoExperiment}
\alias{coerce,}
\alias{SummarizedExperiment,}
\alias{PhosphoExperiment-method}
\title{The PhosphoExperiment class}
\usage{
PhosphoExperiment(
  ...,
  UniprotID = c(),
  GeneSymbol = c(),
  Site = c(),
  Residue = c(),
  Sequence = c(),
  Localisation = c()
)
}
\description{
The PhosphoExperiment class
}
\examples{
library(PhosR)
quant <- as.matrix(PhosR::phospho.L6.ratio)
uniprot <- as.character(sapply(strsplit(rownames(quant),";"), function(x) x[[2]]))
symbol <- as.character(sapply(strsplit(rownames(quant),";"), function(x) x[[2]]))
site <- as.numeric(gsub("[STY]","",sapply(strsplit(rownames(quant),";"), function(x) x[[3]])))
res <- as.character(gsub("[0-9]","",sapply(strsplit(rownames(quant),";"), function(x) x[[3]])))
seq <- as.character(sapply(strsplit(rownames(quant),";"), function(x) x[[4]]))
phosData <- PhosphoExperiment(assays = list(Quantification = quant), UniprotID = uniprot)

}
