% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotKinaseNetwork}
\alias{plotKinaseNetwork}
\title{Plot kinase network}
\usage{
plotKinaseNetwork(KSR, predMatrix, threshold=0.9, color)
}
\arguments{
\item{KSR}{kinase-substrate relationship scoring results}

\item{predMatrix}{output of kinaseSubstratePred function}

\item{threshold}{threshold used to select interconnected kinases for
the expanded signalomes}
}
\value{
a graphical plot
}
\description{
Plot kinase network
}
